+++ pages/posts/[slug].tsx
@@ -14,10 +14,10 @@ import Post from 'entity/post/type'
 import { getFileSlug, getPost, getPostsFilePaths } from 'entity/post/util'
 
 import { Layout, PostTemplate } from 'components'
-import h1 from 'components/mdx/h11'
+import h1 from 'components/mdx/h1'
 import CodeBlock from 'components/mdx/CodeBlock'
-import h2 from 'components/mdx/h22'
-import li from 'components/mdx/lii'
+import h2 from 'components/mdx/h2'
+import li from 'components/mdx/li'
 import p from 'components/mdx/p'
 import ul from 'components/mdx/ul'
 import blockquote from 'components/mdx/blockquote'
